// src/app/models/User/User.model.ts
export interface UserModel {
    id:string;
    userName: string;
    name: string;
    email: string;
    phoneNo: string;
    mobil: string;
  password:string;
    gender: number;
    nationalId: number;
    isActive:number;
    sortable: boolean;
    roleId:string;
    roleName:string;

employeeId:string;
  
  }
  